+++ src/functions.cpp
@@ -5,10 +5,5 @@ float lerp(float min, float max, float blend) {
 }
 
 LevelRect getRect(LevelPos pos, LevelSize size) {
-    float x = pos.x;
-    float y = pos.y;
-    float w = size.w;
-    float h = size.h;
-
-    return { x - w / 2, y - h / 2, w, h };
+    return { pos.x, pos.y, size.w, size.h };
 }