+++ CHANGES.md
@@ -1,8 +1,13 @@
+## HB-1.0.3
+#### ADDED
+- bullet collision with walls
+- bullet rebound
+
 ## HB-1.0.2
 #### ADDED
 - shoot delay
 - weapons
-- wall collisions
+- player collisions with walls and holes
 
 ## HB-1.0.1
 #### ADDED